\name{anorexia.sub}
\alias{anorexia.sub}
\docType{data}
\title{Family Treatment Weight change data for young female anorexia patients.}
\description{
  The MASS package includes the dataset \code{anorexia}, containing pre and post treatment weights for young female anorexia patients.  This is a subset of those data, containing only those patients who received Family Treatment.
}
\usage{data(anorexia.sub)}
\format{A dataframe with 17 observations on the following 2 variables, no NAs.
\describe{
    \item{\code{Prewt}}{Pretreatment weight of subject, in pounds.}
    \item{\code{Postwt}}{Postreatment weight of subject, in pounds.}
 
}}
\source{Hand, D. J., Daly, F., McConway, K., Lunn, D. and Ostrowski, E. eds (1993) A Handbook of Small Data Sets. Chapman & Hall, Data set 285 (p. 229)}

\references{Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.}
\keyword{datasets}
