% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granovaGG-package.R
\docType{data}
\name{anorexia}
\alias{anorexia}
\title{Anorexia Data on Weight Change}
\format{A dataframe with 72 observations of three variables:

\describe{

\item{Treat}{Factor of three levels: "\code{Cont}" (control), "\code{CBT}" (Cognitive Behavioural treatment) and "\code{FT}" (family treatment).}

\item{Prewt}{Pretreatment weight of subject, in pounds.}

\item{Postwt}{Postreatment weight of subject, in pounds.}
}}
\source{
Hand, D. J., Daly, F., McConway, K., Lunn, D. and Ostrowski, E. eds
  (1993) A Handbook of Small Data Sets. Chapman & Hall, Data set 285 (p.
  229)
}
\description{
The anorexia data frame has 72 rows and 3 columns. Weight change data for young female anorexia patients.
}
\references{
Venables, W. N. and Ripley, B. D. (2002) Modern Applied
  Statistics with S. Fourth edition. Springer.
}
\keyword{datasets}

