\docType{data}
\name{arousal}
\alias{arousal}
\title{Arousal in Rats}
\format{A data frame with 40 observations, 10 in each of 4 columns the
corresponding to placebo, drug A, drug B and both drug A and drug B; no
NAs.

\describe{

\item{list("Placebo")}{Rats receiving a placebo treatment.}

\item{list("Drug.A")}{Rats receiving only drug A.}

\item{list("Drug.B")}{Rats receiving only drug B.}

\item{list("Drug.A.B")}{Rats receiving both drug A and drug B.}

}}
\source{
  Richard Lowry. Concepts & Applications of Inferential
  Statistics. Vassar College, Poughkeepsie, N.Y., 2010,
  http://faculty.vassar.edu/lowry/webtext.html
}
\description{
  40 rats were given divided randomly into four groups and
  assigned to one of four treatments: placebo, drug A, drug
  B, or both drug A and drug B. Response is a standard
  measure of physiological arousal.
}
\keyword{datasets}

