\name{blood_lead}
\alias{blood_lead}
\docType{data}
\title{Blood lead levels of lead workers' children matched with similar control children.}
\description{
  Children of parents who had worked in a factory where lead was used in making batteries were matched by age, exposure to traffic, and neighborhood with children whose parents did not work in lead-related industries. Whole blood was assessed for lead content yielding measurements in mg/dl}
\usage{data(anorexia.sub)}
\format{A dataframe with 33 observations on the following 2 variables, no NAs.
\describe{
    \item{\code{Exposed}}{Blood lead level of exposed child, mg/dl.}
    \item{\code{Control}}{Blood lead level of exposed child, mg/dl.}
 
}}
\source{Morton, D., Saah, A., Silberg, S., Owens, W., Roberts, M., Saah, M. (1982). Lead absorption in children of employees in a lead related industry. American Journal of Epidemiology, 115:549-555.}

\references{See discussion in Section 2.5 of Enhancing Dependent Sample Analyses with Graphics, Journal of Statistics Education Volume 17, Number 1 (March 2009).}

\keyword{datasets}
