\docType{data}
\name{rat}
\alias{rat}
\title{Weight gains of rats fed different diets}
\format{A data frame with 60 observations on the following 3 variables, no
  NAs.

\describe{

\item{Weight.Gain}{Weight gain (grams) of rats fed the diets.}

\item{Diet.Amount}{Amount of protein in diet: 1 = High, 2 = Low.}

\item{Diet.Type}{Type of protein in diet: 1 = Beef, 2 = Pork, 3 =
  Cereal.}

}}
\source{
  Fundamentals of Exploratory Analysis of Variance, Hoaglin
  D., Mosteller F. and Tukey J. eds., Wiley, 1991, p. 100;
  originally from Statistical Methods, 7th ed, Snedecor G.
  and Cochran W. (1980), Iowa State Press.
}
\description{
  60 rats were fed varying diets to see which produced the
  greatest weight gain.  Two diet factors were protein
  type: beef, pork, chicken and protein level: high and
  low.
}
\keyword{datasets}

