% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granovaGG-package.R
\docType{data}
\name{shoes}
\alias{shoes}
\title{Shoe wear data of Box, Hunter and Hunter}
\source{
G. E. P. Box, W. G. Hunter and J. S. Hunter (1978) \emph{Statistics for Experimenters}. Wiley, p. 100
}
\description{
A list of two vectors, giving the wear of shoes of materials A and B for one foot each of ten boys.
}
\references{
Venables, W. N. and Ripley, B. D. (2002) \emph{Modern Applied Statistics with S}. Fourth edition. Springer.
}

