% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granovaGG-package.R
\docType{data}
\name{tobacco}
\alias{tobacco}
\title{Virus Preparation on Tobacco Leaves}
\format{
A dataframe with 8 observations on the following 2 variables, no NAs

\describe{
   \item{prep1}{Virus Preparation 1}
   \item{prep2}{Virus Preparation 2}
}
}
\source{
Youden, W. J., Beale, H. P. (1934). A statistical study of the local
  lesion method for estimating tobacco mosaic virus. In Contributions from
  Boyce Thompson Institute 6, page 437.
}
\description{
This data is taken from Snedecor and Cochran (1980) and corresponds to a true
matched pairs experiment. The data originally came from Youden and Beale in
1934 who "wished to find out if two preparations of a virus would produce
different effects on tobacco plants. Half a leaf of a tobacco plant was
rubbed with cheesecloth soaked in one preparation of the virus extract, and
the second half was rubbed similarly with the second extract." (Page 86,
Snedecor and Cochran, 1980) Each of the 8 points in the figure corresponds to
the numbers of lesions on the two halves of one leaf with sides that had been
treated differently.
}
\references{
Snedecor, W., Cochran, W. (1980). Statistical methods. Iowa State
  University Press, Ames Iowa, seventh edition.
}
\keyword{datasets}
