% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granovaGG-package.R
\docType{data}
\name{anorexia.sub}
\alias{anorexia.sub}
\title{Family Treatment Weight change data for young female anorexia patients (subset).}
\format{
A dataframe with 17 observations on the following 2 variables, no
  NAs.

\describe{

\item{\code{Prewt}}{Pretreatment weight of subject, in pounds.}

\item{\code{Postwt}}{Postreatment weight of subject, in pounds.}

}
}
\source{
Hand, D. J., Daly, F., McConway, K., Lunn, D. and Ostrowski, E. eds
  (1993) A Handbook of Small Data Sets. Chapman & Hall, Data set 285 (p.
  229)
}
\description{
The MASS package includes the dataset \code{anorexia}, containing pre and
post treatment weights for young female anorexia patients.  This is a subset
of those data, containing only those patients who received Family Treatment.
}
\references{
Venables, W. N. and Ripley, B. D. (2002) Modern Applied
  Statistics with S. Fourth edition. Springer.
}
\keyword{datasets}
