% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granovaGG-package.R
\docType{data}
\name{arousal}
\alias{arousal}
\title{Arousal in Rats}
\format{A data frame with 40 observations, 10 in each of 4 columns the
  corresponding to placebo, drug A, drug B and both drug A and drug B; no
  NAs.

\describe{

\item{Placebo}{Rats receiving a placebo treatment.}

\item{Drug.A}{Rats receiving only drug A.}

\item{Drug.B}{Rats receiving only drug B.}

\item{Drug.A.B}{Rats receiving both drug A and drug B.}

}}
\source{
Richard Lowry. Concepts & Applications of Inferential Statistics.
  Vassar College, Poughkeepsie, N.Y., 2010,
  http://faculty.vassar.edu/lowry/webtext.html
}
\description{
40 rats were given divided randomly into four groups and assigned to one of
four treatments: placebo, drug A, drug B, or both drug A and drug B.
Response is a standard measure of physiological arousal.
}
\keyword{datasets}

