\name{arousal}
\alias{arousal}
\docType{data}
\title{Arousal in Rats}
\description{
  40 rats were given divided randomly into four groups and assigned to one of four treatments: placebo, drug A, drug B, or both drug A and drug B.  Response is a standard measure of physiological arousal. 
}
\usage{data(arousal)}
\format{A data frame with 40 observations, 10 in each of 4 columns the corresponding to placebo, drug A, drug B and both drug A and drug B; no NAs.
\describe{
    \item{\code{Placebo}}{Rats receiving a placebo treatment.}
    \item{\code{Drug.A}}{Rats receiving only drug A.}
    \item{\code{Drug.B}}{Rats receiving only drug B.}
    
    \item{\code{Drug.A.B}}{Rats receiving both drug A and drug B.}
}}
\source{Richard Lowry. Concepts & Applications of Inferential Statistics. Vassar College, Poughkeepsie, N.Y., 2010,
	 http://faculty.vassar.edu/lowry/webtext.html}
\keyword{datasets}
