\name{GeomRugAlt}
\alias{GeomRugAlt}
\title{Alternate Marginal Rug Plot Geom for ggplot2}
\description{
  This alternate rug plot geom simply plots the rug tufts
  on the top and right (instead of the bottom and left)
}
\author{
  Brian A. Danielak \email{brian@briandk.com}\cr Robert M.
  Pruzek \email{RMPruzek@yahoo.com}

  with contributions by:\cr William E. J. Doane
  \email{wil@drdoane.com}\cr James E. Helmreich
  \email{James.Helmreich@Marist.edu}\cr Jason Bryer
  \email{jason@bryer.org}
}
\references{
  Wickham, H. (2009). Ggplot2: Elegant Graphics for Data
  Analysis. New York: Springer.

  Wilkinson, L. (1999). The Grammar of Graphics. Statistics
  and computing. New York: Springer.
}
\seealso{
  \code{\link{geom_rug}}
}

